fix: dashboard tour as a component
