+++ pages/dashboard/mentor.page.js
@@ -13,9 +13,7 @@ import { useRouter } from "next/router";
 import PreLoader from "../../components/mentorDashboard/PreLoader";
 import { ToastContainer } from "react-toastify";
 import "react-toastify/dist/ReactToastify.css";
-import Image from "next/image";
-import { useTour } from "@reactour/tour";
-import logo from "../../public/logo.png";
+import { MentorDashboardTour } from "./dashboardTour";
 
 function MentorDashboard() {
   // loading and error state
@@ -53,119 +51,7 @@ function MentorDashboard() {
     }
   }, [window.location.search, history]);
 
-  //guided tour
-
-  const { setIsOpen, setSteps, setCurrentStep } = useTour();
-
-  useEffect(() => {
-    setCurrentStep(0);
-    const mentorTour = localStorage.getItem("mentorTour");
-    if (mentorTour) return;
-    setIsOpen(true);
-
-    setSteps([
-      {
-        selector: "#mentorHome",
-        content: (
-          <div className="tw-flex tw-flex-col tw-items-center tw-justify-center">
-            <div className="tw-flex tw-justify-center tw-items-center">
-              <Image src={logo} width={60} height={60} />
-            </div>
-            <p className="tw-text-black">Welcome to the mentor dashboard</p>
-            <div className="tw-flex tw-justify-center tw-items-center">
-              <span className="tw-font-bold tw-text-xl">GrabTern</span>
-            </div>
-          </div>
-        ),
-      },
-      {
-        selector: "#mentorSidebar",
-        content:
-          "This is the sidebar. You can navigate to other pages from here.",
-      },
-      {
-        selector: "#publishProfile",
-        content: "You can publish or unpublish your profile from here.",
-      },
-      {
-        selector: "#mentorSidebarToggle",
-        content: "You can toggle the sidebar from here.",
-      },
-      {
-        selector: "#mentorProfile",
-        content: "This is your profile. You can edit your profile from here.",
-      },
-      {
-        selector: "#mentorHomepage",
-        content:
-          "This is your homepage. You can view your bookings and sessions from here.",
-      },
-      {
-        selector: "#mentorBookings",
-        content: "This is your bookings. You can view your bookings from here.",
-      },
-      {
-        selector: "#mentorSessions",
-        content: "This is your sessions. You can view your sessions from here.",
-      },
-      {
-        selector: "#mentorQueries",
-        content: "This is your queries. You can view your queries from here.",
-      },
-      {
-        selector: "#mentorCalendar",
-        content: "This is your calendar. You can view your sessions from here.",
-      },
-      {
-        selector: "#mentorServices",
-        content: "This is your services. You can view your services from here.",
-      },
-      {
-        selector: "#mentorPayments",
-        content:
-          "This is your payments. You can view and manage your payments from here.",
-      },
-      {
-        selector: "#mentorWhatsNew",
-        content:
-          "This is your what's new. You can view your what's new from here.",
-      },
-      {
-        selector: "#mentorReferral",
-        content:
-          "This is your referral. You can view your referrals from here, Refer and Earn.",
-      },
-      {
-        selector: "#mentorRewards",
-        content: "This is your rewards. You can view your rewards from here.",
-      },
-      {
-        selector: "#mentorNotifications",
-        content:
-          "This is your notifications. You can view your notifications from here.",
-      },
-      {
-        selector: "#mentorLogout",
-        content: "This is your logout. You can logout from here.",
-      },
-      {
-        selector: "#mentorHome",
-        content: (
-          <div className="tw-flex tw-flex-col tw-items-center tw-justify-center">
-            <div className="tw-flex tw-justify-center tw-items-center">
-              <Image src={logo} width={60} height={60} />
-            </div>
-            <p className="tw-text-black">Have a great time</p>
-            <div className="tw-flex tw-justify-center tw-items-center">
-              <span className="tw-font-bold tw-text-xl">GrabTern</span>
-            </div>
-          </div>
-        ),
-      },
-    ]);
-
-    localStorage.setItem("mentorTour", true);
-  }, []);
+  MentorDashboardTour();
 
   if (!isMentorLoggedIn) {
     return null;
